fix: renamed test files, to be picked up and run
